+++ webapp/src/app/lookup_pk/[wallet]/route.ts
@@ -1,10 +1,12 @@
 import { sql } from "@vercel/postgres";
 import { NextRequest } from "next/server";
+import { unstable_noStore } from "next/cache";
 
 export async function GET(
   request: NextRequest,
   { params }: { params: { wallet: string } }
 ) {
+  unstable_noStore();
   const { rows } =
     await sql`SELECT nfc_pk from ff_users where embedded_wallet=${params.wallet}`;
