+++ webapp/src/app/lookup_wallet/[publicKey]/route.ts
@@ -1,10 +1,12 @@
 import { sql } from "@vercel/postgres";
+import { unstable_noStore } from "next/cache";
 import { NextRequest } from "next/server";
 
 export async function GET(
   request: NextRequest,
   { params }: { params: { publicKey: string } }
 ) {
+  unstable_noStore();
   const { rows } =
     await sql`SELECT embedded_wallet from ff_users where nfc_pk =${params.publicKey}`;
